Add leader election to avoid schema disagreements with concurrent migrations
When starting multiple distributed processes at the same time, migrations will be executed at the same time, often resulting in schema conflicts.
This commit adds leader election through LWTs so that only a single process will perform the migration at once on a given keyspace.
Leader election will be optional and deactivated by default.
`IF NOT EXISTS` clauses were added to the `CREATE TABLE` statements in order to avoid conflicts there as well.
Both changes are disabled if the Cassandra version is below 2.0 since we would miss both LWTs and `IF NOT EXISTS` features.